// src/utils/trackTranslationOverrides.generated.ts
export interface TrackTranslationOverrideRowGen {
  requestTypeName: string;
  providerName: string;
  defaultTranslation: string;
  customTranslation: string;
  trackCode: number;
}